pyproject.toml

[project]
name = ""
version = ""
description = ""
authors = [
    {name = "eliaswhite", email = "souzabrancoelias8@gmail.com"},
]
dependencies = [
    "django>=4.2.4",
    "djangorestframework>=3.14.0",
    "setuptools>=68.1.0",
]
requires-python = ">=3.10"
license = {text = "MIT"}
